refactor: move ArrowLeft into Information.tsx
Added a component property to walkThroughSteps, so now we can pass
different components, eg ArrowLeft, ArrowRight, etc. Added a nav-height
variable to GlobalStyles. We can use this to set the nav height AND
subtract the nav height when sizing elements that need to be 100vh minus
the nav height. Information.tsx now renders the component to highlight
what the modal is talking about (eg a button). As we cycle through the
steps, the appropriate component is shown.