refactor: move logic out of Grid into Information
We can create an array of refs that refer to our steps in Grid.tsx.
The can be passed to Information, and then used to calculate the
position to place our component (arrow up, arrow left etc) for our
current step. Fixed issue in Cell.tsx where red border was not showing
up on invalid cells due to !important when setting black border. Added
stepRefs, array of refs with useRef, to Grid.tsx. Passed this as a prop
to Information.tsx.