feat: add code for tutorial feature
Renamed ArrowLeftProps to PositionProps, I'll want to use them with
different components. Added right, bottom, height, and width props to
PositionProps. We pass the height prop into our styled ArrowLeft
component and use this to set the top position. This allows us to center
our arrow or component on the element the arrow is referring to. Removed
code from Grid.tsx that related to position and the useEffect to set the
position state. This logic has been moved to Information.tsx. Added a
ref for each element in Grid.tsx that has a corresponding step. Added
a class name to each of these components that matches the step id in
walkThroughSteps. Added a Side enum. We may want to attach our step
component to a particular side of the app component for a given step.
For example, an UpArrow component would want be attached to the bottom
of an app component that it is pointing towards. In walkThroughSteps,
made the component prop optional and added an attach propery that uses
our new Side enum, so Side.TOP etc.