feat: position tutorial components dynamically
Renamed ArrowLeft to Arrow to reflect its more generalized usage.
Arrow now imports up, down, left and right arrows from react icons.
Added the side prop to Arrow. This comes from the optional attach prop
of walkthroughSteps. The side prop specifices which side of a component
in our app that our Arrow should attach to. Added a switch statement
to Arrow for our 4 arrow types. This allows us to set the imported icon
to use and calculate its top, left, transformX and transformY vals that
are then passed to our styled component. Example: side is Side.TOP. We
use the arrow down component, and align the bottom edge of arrow down
with the top edge of the component it is attached to. We align it to the
middle of the component that it is attached to.