feat: add modal to warn user of invalid grid
Added Modal.tsx to display warnings to the user when the grid is in an
invalid state. Slight refactor to useModal hook so it accepts initial
state for isVisible - true or false. Added a ref and another useEffect
to Grid.tsx. The ref stores the previous value of the isValid state.
Using this, we can ensure that the warning is only shown when the grid
state switches from a valid state to an invalid state and not when the
isValid state remains the same between rerenders.