feat: add simple warning to user if grid invalid
Decided that a modal each time the isValid state switches form true to
false would be incredibly annoying. Replaced this with a warning that
appears above the crossword grid with a pulsing information icon.
Clicking the icon opens the modal, where we can explain to the user what
the issue is. Added animation/keyframes for the icon. Added styling
for the warning.