refactor: each step can render a component
Instead of rendering the contents of a 'text' prop in a step, each step
now has a 'content' prop that is a React.FC or a ReactNode. I was going
to also allow strings but decided to avoid using
dangerouslySetInnerHTML. Added a TooShort component as the content for
the first step of invalidGridSteps.ts. Updated Information.tsx to render
our steps using the updated model/steps. Added an EmptyProps type to
avoid using {} as a type (TS complaint).