feat: finish steps for invalid grid + styling
Finished the last steps for invalidGrid. Fixed styling for Information
so that there is more room between the edge of the modal and the
content. Fixed the overflow on the prevent-click overlay in Grid.tsx by
setting its container to position relative. Adjusted absolute position
of the invalid grid display to ensure it is always outside of the grid
container. In Information: set the top position and height so that
Information covers the entire screen including the navbar. Removed the
offset for the navbar from Arrow.tsx. Fixed issue in Root.tsx where
the OpenAI logo was not lining up with the text "Powered By OpenAI."