refactor: if there are no clues, display message
In SolveGrid.tsx, if there are no clues in local storage or state,
instead of displaying clues with no content, we display a message to
let the user know that clues will be displayed in this area once they
generate a new puzzle or select a puzzle from the dropdown. Update
styles so this is displayed correctly on phone and large screens.
Removed a TypeScript error for the error prop passed to ErrorPage.
Added an interface for ErrorPage - a bit hacky.