feat: when user selects an answer, it is displayed
On mobile screens, the clues are shown below he grid and may be
offscreen or hidden by the keyboard. Added a renderSelectedClue function
to SolveGrid that displays the clue for the selected answer on the grid.
This is displayed above the crossword grid, so the user does not have to
scroll down to see what the clue is. Adjusted styling for check, reveal,
and clear buttons on mobile - larger font, spaced further apart. Updated
some other styles for responsiveness/better layout on mobile.