refactor: change elements used to display clues
On small screens, when the user clicks on the grid, we show the clue
relating to their selection. This clue and our other clues are now
rendered in divs rather than list item elements. The primary reason for
the refactor is so clue text wraps on smaller screens and we have a gap
between the clue number, the clue text, and the length of the answer.
Updated all max-width: 500 media queries to be 600, and updated our
min-width: 501 to be 601.