refactor: add copy of SolveCell.tsx, update notes
Added notes relating to key strokes not working on mobile. This is due
to e.key being "Unidentified" on my Samsung phone. This will require
a big refactor in how input is processed in the app. As the app
currently stands, the value of e.key drove the logic of what happened
after a user pressed a key. See NOTES.md for more detail.