refactor: make tab, backspace & alpha keys work
Tab now partially works as expected on physical keyboards. You can tab
forwards and backwards between answers/clues but when cycling backwards
you will break out of the grid and tab to other elements on the page.
The 2nd cell of the answer is being focused - might require prevent
default. Alpha keys are working as expected. Backspace is working as
expected.