refactor: add code to handle lack of e.key mobile
Added e.preventDefault to handleTabKeyPress - resolved issue where
cycling backwards through clues/answers was causing other controls to
be focused. Fixed issue where the user could not select a cell and type
into that cell thus overriding the existing content. This involves
allowing the user to type more than 1 character into an input and
setting the state to be the last character typed.