fix: rename SolveGridRefactor, delete getCluesOLD
Renamed SolveGridRefactor to SolveGrid. In Grid.tsx, check if the grid
has been updated from the initial state by stringifying the initialState
grid object and comparing it to the stringified gridState. Changed logic
for displaying Information.tsx for our onboarding steps so that the
modal is not shown if there is local storage, i.e. only show on first
visit. In populateClues in utils.ts, replaced a forEach with a for...of.