refactor: add GridRefactor
Added GridRefactor.tsx in attempt to use our utilsRefactor in our Grid
component. Added code in generateClues of Grid.tsx to repeatedly call
populateClues while the grid state was invalid. Added code to
populateClues in utils and utilsRefactor after empty cells are removed.
We check if the grid is valid, and if it is not we attempt to reset
single island cells removing them from the grid. We then revalidate
the grid again. Fixed some errors in utilsRefactor.