fix: fix infinite loop and no grid displaying
Fixed issue where no grid was being displayed. The generateClues func
was inside a setTimeout. The state setter was outside the time out and
so we were setting the state to an empty array. The optional chaining
for rendering meant that if there is no state then nothing gets rendered.
Fixed issue where a while loop is regenerating the clues/grid while the
grid is invalid. The issue was failing to update the while condition in
the body of the loop!