feat: add func to detect blocks of cells
Added findBlockOfCells function that detects blocks of light cells.
Since our code won't correctly generate answers for a 2x2 or 3x3 or
NxN block of cells, any block of such cells will cause the grid to be
in an invalid state. The function returns an isValid variable that is
either true or false. It also sets the isValid prop on each cell that is
part of the 2x2 block to be false.