feat: add step to invalid grid modal
Added a step to invalidGridSteps.ts for situations where we have a block
of contiguous answers. Our code will generate nonsense answers if the
grid is not a lattice/there is a 2x2 or greater block of cells next to
each other. Setting this grid state as invalid will prevent the user
from fetching clues from OpenAI in the create route and potentially end
up trying to solve a puzzle with non-words.